requirements_streamlit.txt
streamlit>=1.32.0
pandas>=2.0.0
openpyxl>=3.1.0
rapidfuzz>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
